add alt colour modes
they work ok with a sine-wave tone generator, and pretty badly with basic vocal
testing.

either a) everything's working well, and vocals are just pretty complex,
crunching the colour changes highlights this and looks bad, or b) the frequency
detection isn't actually that good and leaks a lot.

It's probably (and hopefully) b), since the implemntation is pretty hack.
At the least because there's no windowing on the samples.

On the other side of that, each goertzel iteration runs on the full 2048 sample
window currently, which is
- wasteful
- adds latency to higher frequencies
- makes the bins extra narrow

It's not clear to me whether the narrow bins are helping or hurting at the
moment, but it's probably time to start graphing and measuring better.